Fix heap corruption caused by specifying an x value outside of display
Bug:
The memory corruption occurs during the memset call in the erase
routine.  When bad values are in drawBitmap the number of bytes to zero
is too long by 1 byte.  These values are updated by several macros, one
of which is pageCheckBoundsRange.  The values passed to this macro
should be x values that match physical pixels on the screen.  However
calls to drawBitmap was using the number of screen pixels or width to
offset the current x value resulting in the next pixel (not updated)
instead of the last pixel (updated).  When the bitmap or a portion of
the bitmap is at the edge of the screen, this results in an invalid x
value outside the range of (0 <= x < width).

Fix:
* Compute the last updated pixel offset for the pageCheckBoundsRange in
drawBitmap
* Add comments to pageCheckBounds* macros to specify bounds in hopes of
avoiding a similar issue in the future